backend/requirements.txt
asgiref==3.5.2
Django==3.2.15
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
numpy==1.23.3
opencv-python==4.6.0.66
pickle-mixin==1.0.2
Pillow==9.2.0
python-decouple==3.6
pytz==2022.2.1
sockets==1.0.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
